+++ NEWS.md
@@ -7,3 +7,11 @@ ViSiElse 1.2.1 (R package) is a visual tool for behaviour analysis.
 * Replace deprecated function Matrix::rBind 
 
 * Fix plotting statistical mean and standard deviations issues
+
+## ViSiElse paper
+
+* Intubation example from the "ViSiElse: An innovative visualization R package 
+to ensure behavioral raw data transparency and reliability" paper
+
+* Visit the Github /Example repository to find the intubation data example
+(ViSiElse_intubation_data.csv) and the associated R code (ViSiElse_intubation_R_code.R)